move setup call to settings screen
