+++ res/menu/drawer_menu.xml
@@ -63,11 +63,6 @@
     <group android:id="@+id/drawer_menu_bottom" android:checkableBehavior="single">
         <item
             android:orderInCategory="3"
-            android:id="@+id/nav_dav_client"
-            android:icon="@drawable/ic_launch_grey"
-            android:title="@string/drawer_calendar_contacts"/>
-        <item
-            android:orderInCategory="3"
             android:id="@+id/nav_settings"
             android:icon="@drawable/ic_settings"
             android:title="@string/actionbar_settings"/>
