+++ res/values/setup.xml
@@ -56,6 +56,7 @@
     <bool name="multiaccount_support">true</bool>
     
     <!-- Help, imprint and feedback -->
+    <bool name="calendar_contacts_enabled">true</bool>
     <bool name="help_enabled">true</bool>
     <bool name="imprint_enabled">false</bool> 
     <bool name="recommend_enabled">true</bool>
