+++ res/xml/preferences.xml
@@ -78,6 +78,9 @@
 	</PreferenceCategory>
 
 	<PreferenceCategory android:title="@string/prefs_category_more" android:key="more">
+		<Preference android:title="@string/prefs_calendar_contacts"
+					android:key="calendar_contacts"
+					android:summary="@string/prefs_calendar_contacts_summary" />
 		<Preference android:title="@string/prefs_help" android:key="help" />
 		<Preference android:title="@string/prefs_recommend" android:key="recommend" />
 		<Preference android:title="@string/prefs_feedback" android:key="feedback" />
@@ -85,6 +88,5 @@
 		<Preference android:title="@string/prefs_imprint" android:key="imprint" />
 		<Preference android:title="@string/about_title" android:id="@+id/about_app" android:key="about_app" />
 	</PreferenceCategory>
-
 
 </PreferenceScreen>
